[WOR-431] Refactor method to companion object on ServicePerimeterService
